requirements.txt
kafka-python==2.0.2
playsound==1.3.0
